Debugging case where it is not possible to reach goal
An if condition is set at the begining of the repairing process since there may be a case where there are no Global Waypoints in the trajectory. This could be the reason why an error occurs after declaring the goal is in forbidden area.
Besides, two output texts are also included to check what happens when the index variable (which is an index pointing to the Global Waypoint in globalPath to reconnect) exceeds the number of Global Waypoints.